Restructure resolvers directory
This will make it a lot easier to scale and review. Awesome!
