+++ src/resolvers/mutations/users.ts
@@ -1,0 +1,53 @@
+import {Users} from '@app/models.js'
+import {Role} from '@app/types/generated/schema.js'
+import {authenticateUser} from '@helpers/authentication.js'
+import {authorizeRoleAccess, authorizeOwnership} from '@helpers/authorization.js'
+import {signJWT} from '@helpers/sign-jwt.js'
+import bcrypt from 'bcrypt'
+import type {MutationResolvers as Resolvers} from '@app/types/generated/schema.js'
+
+const createUser: Resolvers['createUser'] = async (
+	parent,
+	args,
+	context,
+) => {
+	const user = await Users.create({
+		email: args.email.trim( ),
+		username: args.username.trim( ),
+		password: await bcrypt.hash(args.password, 12),
+	})
+
+	return signJWT(user)
+}
+
+const signInUser: Resolvers['signInUser'] = async (
+	parent,
+	args,
+	context,
+) => {
+	// Check username & password, and get info via username if good-to-go.
+	await authenticateUser(args.username, args.password)
+	const user = await Users.findOne({username: args.username})
+
+	return signJWT(user)
+}
+
+const deleteUser: Resolvers['deleteUser'] = async (
+	parent,
+	args,
+	context,
+) => {
+	await Promise.any([
+		authorizeOwnership(context.user, args.id),
+		authorizeRoleAccess(context.user, Role.OWNER),
+	])
+
+	const user = await Users.deleteOne({_id: args.id})
+	return user.acknowledged
+}
+
+export {
+	createUser,
+	signInUser,
+	deleteUser,
+}
