+++ src/resolvers/objects/index.ts
@@ -1,0 +1,3 @@
+export * from './course.js'
+export * from './project.js'
+export * from './user.js'
