+++ src/resolvers/queries/courses.ts
@@ -1,0 +1,23 @@
+import {Courses} from '@app/models.js'
+import type {QueryResolvers as Resolvers} from '@app/types/generated/schema.js'
+
+const getCourse: Resolvers['getCourse'] = async (
+	parent,
+	args,
+	context,
+) => (
+	await Courses.findOne({id: args.id}) ?? null
+)
+
+const indexCourses: Resolvers['indexCourses'] = async (
+	parent,
+	args,
+	context,
+) => (
+	await Courses.find({ })
+)
+
+export {
+	getCourse,
+	indexCourses,
+}
