+++ src/resolvers/queries/users.ts
@@ -1,0 +1,13 @@
+import type {QueryResolvers as Resolvers} from '@app/types/generated/schema.js'
+
+const getMe: Resolvers['getMe'] = (
+	parent,
+	args,
+	context,
+) => (
+	context.user ?? null
+)
+
+export {
+	getMe,
+}
